requirements.txt
cachelib==0.9.0
click==8.1.3
configparser==5.3.0
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-Session==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
PyMySQL==1.0.2
Werkzeug==2.2.2
